Start refactoring UserAuthEndpointer to support host auth
