+++ server/service/osquery.go
@@ -1,0 +1,189 @@
+package service
+
+import (
+	"context"
+	"encoding/json"
+
+	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
+	hostctx "github.com/fleetdm/fleet/v4/server/contexts/host"
+	"github.com/fleetdm/fleet/v4/server/fleet"
+	"github.com/spf13/cast"
+)
+
+////////////////////////////////////////////////////////////////////////////////
+// Get Client Config
+////////////////////////////////////////////////////////////////////////////////
+
+type getClientConfigRequest struct {
+	NodeKey string `json:"node_key"`
+}
+
+type getClientConfigResponse struct {
+	Config map[string]interface{}
+	Err    error `json:"error,omitempty"`
+}
+
+func (r getClientConfigResponse) error() error { return r.Err }
+
+func getClientConfigEndpoint(ctx context.Context, request interface{}, svc fleet.Service) (interface{}, error) {
+	config, err := svc.GetClientConfig(ctx)
+	if err != nil {
+		return getClientConfigResponse{Err: err}, nil
+	}
+
+	// We return the config here explicitly because osquery exepects the
+	// response for configs to be at the top-level of the JSON response
+	return config, nil
+}
+
+func (svc *Service) GetClientConfig(ctx context.Context) (map[string]interface{}, error) {
+	// skipauth: Authorization is currently for user endpoints only.
+	svc.authz.SkipAuthorization(ctx)
+
+	host, ok := hostctx.FromContext(ctx)
+	if !ok {
+		return nil, osqueryError{message: "internal error: missing host from request context"}
+	}
+
+	baseConfig, err := svc.AgentOptionsForHost(ctx, host.TeamID, host.Platform)
+	if err != nil {
+		return nil, osqueryError{message: "internal error: fetch base config: " + err.Error()}
+	}
+
+	config := make(map[string]interface{})
+	if baseConfig != nil {
+		err = json.Unmarshal(baseConfig, &config)
+		if err != nil {
+			return nil, osqueryError{message: "internal error: parse base configuration: " + err.Error()}
+		}
+	}
+
+	packs, err := svc.ds.ListPacksForHost(ctx, host.ID)
+	if err != nil {
+		return nil, osqueryError{message: "database error: " + err.Error()}
+	}
+
+	packConfig := fleet.Packs{}
+	for _, pack := range packs {
+		// first, we must figure out what queries are in this pack
+		queries, err := svc.ds.ListScheduledQueriesInPack(ctx, pack.ID)
+		if err != nil {
+			return nil, osqueryError{message: "database error: " + err.Error()}
+		}
+
+		// the serializable osquery config struct expects content in a
+		// particular format, so we do the conversion here
+		configQueries := fleet.Queries{}
+		for _, query := range queries {
+			queryContent := fleet.QueryContent{
+				Query:    query.Query,
+				Interval: query.Interval,
+				Platform: query.Platform,
+				Version:  query.Version,
+				Removed:  query.Removed,
+				Shard:    query.Shard,
+				Denylist: query.Denylist,
+			}
+
+			if query.Removed != nil {
+				queryContent.Removed = query.Removed
+			}
+
+			if query.Snapshot != nil && *query.Snapshot {
+				queryContent.Snapshot = query.Snapshot
+			}
+
+			configQueries[query.Name] = queryContent
+		}
+
+		// finally, we add the pack to the client config struct with all of
+		// the pack's queries
+		packConfig[pack.Name] = fleet.PackContent{
+			Platform: pack.Platform,
+			Queries:  configQueries,
+		}
+	}
+
+	if len(packConfig) > 0 {
+		packJSON, err := json.Marshal(packConfig)
+		if err != nil {
+			return nil, osqueryError{message: "internal error: marshal pack JSON: " + err.Error()}
+		}
+		config["packs"] = json.RawMessage(packJSON)
+	}
+
+	// Save interval values if they have been updated.
+	intervalsModified := false
+	intervals := fleet.HostOsqueryIntervals{
+		DistributedInterval: host.DistributedInterval,
+		ConfigTLSRefresh:    host.ConfigTLSRefresh,
+		LoggerTLSPeriod:     host.LoggerTLSPeriod,
+	}
+	if options, ok := config["options"].(map[string]interface{}); ok {
+		distributedIntervalVal, ok := options["distributed_interval"]
+		distributedInterval, err := cast.ToUintE(distributedIntervalVal)
+		if ok && err == nil && intervals.DistributedInterval != distributedInterval {
+			intervals.DistributedInterval = distributedInterval
+			intervalsModified = true
+		}
+
+		loggerTLSPeriodVal, ok := options["logger_tls_period"]
+		loggerTLSPeriod, err := cast.ToUintE(loggerTLSPeriodVal)
+		if ok && err == nil && intervals.LoggerTLSPeriod != loggerTLSPeriod {
+			intervals.LoggerTLSPeriod = loggerTLSPeriod
+			intervalsModified = true
+		}
+
+		// Note config_tls_refresh can only be set in the osquery flags (and has
+		// also been deprecated in osquery for quite some time) so is ignored
+		// here.
+		configRefreshVal, ok := options["config_refresh"]
+		configRefresh, err := cast.ToUintE(configRefreshVal)
+		if ok && err == nil && intervals.ConfigTLSRefresh != configRefresh {
+			intervals.ConfigTLSRefresh = configRefresh
+			intervalsModified = true
+		}
+	}
+
+	// We are not doing deferred update host like in other places because the intervals
+	// are not modified often.
+	if intervalsModified {
+		if err := svc.ds.UpdateHostOsqueryIntervals(ctx, host.ID, intervals); err != nil {
+			return nil, osqueryError{message: "internal error: update host intervals: " + err.Error()}
+		}
+	}
+
+	return config, nil
+}
+
+// AgentOptionsForHost gets the agent options for the provided host.
+// The host information should be used for filtering based on team, platform, etc.
+func (svc *Service) AgentOptionsForHost(ctx context.Context, hostTeamID *uint, hostPlatform string) (json.RawMessage, error) {
+	// Team agent options have priority over global options.
+	if hostTeamID != nil {
+		teamAgentOptions, err := svc.ds.TeamAgentOptions(ctx, *hostTeamID)
+		if err != nil {
+			return nil, ctxerr.Wrap(ctx, err, "load team agent options for host")
+		}
+
+		if teamAgentOptions != nil && len(*teamAgentOptions) > 0 {
+			var options fleet.AgentOptions
+			if err := json.Unmarshal(*teamAgentOptions, &options); err != nil {
+				return nil, ctxerr.Wrap(ctx, err, "unmarshal team agent options")
+			}
+			return options.ForPlatform(hostPlatform), nil
+		}
+	}
+	// Otherwise return the appropriate override for global options.
+	appConfig, err := svc.ds.AppConfig(ctx)
+	if err != nil {
+		return nil, ctxerr.Wrap(ctx, err, "load global agent options")
+	}
+	var options fleet.AgentOptions
+	if appConfig.AgentOptions != nil {
+		if err := json.Unmarshal(*appConfig.AgentOptions, &options); err != nil {
+			return nil, ctxerr.Wrap(ctx, err, "unmarshal global agent options")
+		}
+	}
+	return options.ForPlatform(hostPlatform), nil
+}
